WIP: Added concurent version of getDistances func
WARNING: getDistancesNoChan is the concurrent version of getDistances.
It will work for <10 API calls, but it will fail for more. This seems to
be the rate set by osrm. The concurrency itself is sound, and has been tested
by calling this API: https://xkcd.com/22/info.0.json. Significant performance
improvment can be seen. This function needs to be tested with a local osrm
server, in a docker container.